pyproject.toml
[project]
name = "mcpylib-map-art"
version = "0.1.0"
description = "A community tool for creating Minecraft map art using MCPyLib"
readme = "README.md"
license = {text = "Dual License: Free for non-commercial use, commercial use requires authorization"}
requires-python = ">=3.11"
authors = [
    {name = "MCPyLib Map Art Contributors", email = "treeleaves30760@gmail.com"}
]
keywords = ["minecraft", "map-art", "mcpylib", "pixel-art", "image-processing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Games/Entertainment",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcpylib>=1.0.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
]

[project.scripts]
mcpylib-map-art = "map_art.cli:cli"
